Split codebase into modules
