+++ src/utils.js
@@ -1,0 +1,5 @@
+const capitalize = (name) => name.charAt(0).toUpperCase() + name.slice(1);
+
+module.exports = {
+  capitalize,
+};
